requirements.txt
asgiref
autopep8
django==3.0.8
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
pymysql
Pillow
django-datatables-view
djangorestframework-datatables
django-session-timeout
django-debug-toolbar
django-crispy-forms
gtts
playsound
Image
django-imagefit

